Add grid spacing of four to bob template file
There is a mechanic with phoebus group objects where the widgets within them are positioned relative to that grid object. This means that those widgets are not in line with the screens grid that the group object itself (and any top level widgets) are placed on. To overcome this, there are two options: We either reduce the font size to a non-standard 15, or reduce the grid size by 4. This update opts for the latter.