Create a place holder convert for plugins
This provides a convert which takes a plugin from /ADCore/ADApp/pluginSrc/ that has both a .h and .cpp file and makes a minor adjustment in each.

In the .h it adds the parent class paramSet into the class definition. Such as:

class NDPLUGIN_API NDPluginAttribute : public NDPluginDriverParamSet, public NDPluginDriver

In the .cpp it adds "static_cast<{parent}ParamSet*>(this)" to the base class constructor:

NDPluginAttribute::NDPluginAttribute(...)
    : NDPluginDriver(static_cast<NDPluginDriverParamSet*>(this), ...

It produces these file in ADCore/pvi/placeholders