Divide max_x/y functions into max_x/y and next_x/y to be more explicit
Update 'widgets' argument to take in a single WidgetFactory or a list of WidgetFactorys as a union type.
There are times where max_x and max_y were used to calculate the next position that a widget could be placed at, this was confusing. These changes address the dual purpose of these functions by breaking them up into explicit functions.