Add regrouping of producer.yaml pv on .edl files
Adds a regrouping util, a regrouping func in main and launch configs for it. The regrouping works for edl screens of the type in ADCore/ADApp/op/edl, does not work for /autoconvert screens. The rectangle-name matching in the regrouping utill does not function as desired for /autoconvert screens.

Grouping is very sensitive to the format of the edl file and PV's will be regrouped incorrectly if the labelled rectangles are nested within larger ones.