Change widget layout to place in last column
Now new group and widget placement behaves the same: They'll be placed
in the rightmost column if they fit, or into a new column.
This means the order of rendered widgets will exactly match the input
order - we no longer reshuffle in an attempt to preserve screen space.

Previously new groups would be placed into the leftmost column where it
would fit, or placed into the last column.
Standalone widgets would be placed in either of the last two columns, if
it would fit.