Fix SignalRW read_pv convert
The conversion was defaulting to use name + RBV for read_pv,
which is now not correct.

This fix makes SettingPair pass in the read_record so that
the pv comes from the template.

It would probably be possible to refactor these classes so that
more fields are required and checks for None can be removed.